refactor: Tidy up ts files into command dir
