+++ src/adaptor/discord/sheriff.ts
@@ -1,5 +1,5 @@
 import { ChannelType, Client } from 'discord.js';
-import type { Sheriff } from '../../service/stfu.js';
+import type { Sheriff } from '../../service/command/stfu.js';
 import type { Snowflake } from '../../model/id.js';
 
 export class DiscordSheriff implements Sheriff {
