+++ src/adaptor/in-memory-reservation.ts
@@ -2,7 +2,7 @@ import type { Reservation, ReservationTime } from '../model/reservation.js';
 import type {
   ReservationRepository,
   ReservationResult
-} from '../service/kaere.js';
+} from '../service/command/kaere.js';
 
 export class InMemoryReservationRepository implements ReservationRepository {
   map: Map<string, Reservation> = new Map();
