+++ src/server/index.ts
@@ -18,7 +18,7 @@ import { Client, GatewayIntentBits, version } from 'discord.js';
 import type {
   CommandMessage,
   CommandResponder
-} from '../service/command-message.js';
+} from '../service/command/command-message.js';
 import {
   EmojiResponseRunner,
   MessageResponseRunner,
@@ -38,7 +38,7 @@ import {
   allRoleResponder,
   registerAllCommandResponder
 } from '../service/index.js';
-import type { AssetKey } from '../service/party.js';
+import type { AssetKey } from '../service/command/party.js';
 import { DiscordMemberStats } from '../adaptor/discord/member-stats.js';
 import { DiscordMessageRepository } from '../adaptor/discord/message-repo.js';
 import { DiscordRoleManager } from '../adaptor/discord/role.js';
@@ -46,7 +46,7 @@ import { DiscordSheriff } from '../adaptor/discord/sheriff.js';
 import { DiscordWS } from '../adaptor/discord/ws.js';
 import { EmojiProxy } from '../adaptor/emoji-proxy.js';
 import { GenVersionFetcher } from '../adaptor/version/fetch.js';
-import type { KaereMusicKey } from '../service/kaere.js';
+import type { KaereMusicKey } from '../service/command/kaere.js';
 import { Snowflake } from '../model/id.js';
 import dotenv from 'dotenv';
 import { extractEnv } from './extract-env.js';
