+++ src/service/command/debug.test.ts
@@ -1,6 +1,6 @@
 import { DebugCommand, MessageRepository } from './debug.js';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { Snowflake } from '../model/id.js';
+import type { Snowflake } from '../../model/id.js';
 import { createMockMessage } from './command-message.js';
 
 describe('debug', () => {
