+++ src/service/command/meme/lolicon.ts
@@ -1,4 +1,4 @@
-import type { MemeTemplate } from '../../model/meme-template.js';
+import type { MemeTemplate } from '../../../model/meme-template.js';
 
 export const lolicon: MemeTemplate<never, never> = {
   commandNames: ['lolicon'],
