+++ src/service/command/party.test.ts
@@ -1,8 +1,8 @@
 import { type AssetKey, PartyCommand, type RandomGenerator } from './party.js';
-import { MockClock, MockVoiceConnectionFactory } from '../adaptor/index.js';
+import { MockClock, MockVoiceConnectionFactory } from '../../adaptor/index.js';
 import { afterAll, describe, expect, it, vi } from 'vitest';
-import type { EmbedMessage } from '../model/embed-message.js';
-import { ScheduleRunner } from '../runner/index.js';
+import type { EmbedMessage } from '../../model/embed-message.js';
+import { ScheduleRunner } from '../../runner/index.js';
 import { createMockMessage } from './command-message.js';
 
 const random: RandomGenerator = {
