+++ src/service/command/role-rank.test.ts
