+++ src/service/command/role-rank.ts
@@ -3,7 +3,7 @@ import type {
   CommandResponder,
   HelpInfo
 } from './command-message.js';
-import { MessageEvent } from '../runner/message.js';
+import { MessageEvent } from '../../runner/message.js';
 
 export interface MemberWithRole {
   displayName: string;
