+++ src/service/command/typo-record.test.ts
@@ -1,4 +1,4 @@
-import { InMemoryTypoRepository, MockClock } from '../adaptor/index.js';
+import { InMemoryTypoRepository, MockClock } from '../../adaptor/index.js';
 import { SentMessage, createMockMessage } from './command-message.js';
 import {
   TypoRecorder,
@@ -7,10 +7,10 @@ import {
 } from './typo-record.js';
 import { addDays, setHours, setMinutes } from 'date-fns';
 import { afterAll, describe, expect, it, vi } from 'vitest';
-import type { EmbedMessage } from '../model/embed-message.js';
+import type { EmbedMessage } from '../../model/embed-message.js';
 import EventEmitter from 'node:events';
-import { ScheduleRunner } from '../runner/index.js';
-import type { Snowflake } from '../model/id.js';
+import { ScheduleRunner } from '../../runner/index.js';
+import type { Snowflake } from '../../model/id.js';
 
 class MockRepository extends EventEmitter implements TypoRepository {
   private db = new InMemoryTypoRepository();
